Create GithubService.java
Implemented GithubService class for fetching GitHub data

Created GithubService class to encapsulate the logic for fetching user data from the GitHub API. Utilizes RestTemplate for making HTTP requests to the GitHub API endpoints. The getUserData method fetches user data and repositories based on the provided username. It processes the retrieved data and constructs a GithubUser object representing the user and their repositories. Included logic to convert the ISO 8601 timestamp for user creation date to the desired format. This service facilitates the integration of GitHub data into the application.